wolf-of-nyc/requirements.txt

gradio>=4.0.0
gtts>=2.3.0
pytesseract>=0.3.10
Pillow>=10.0.0
setuptools>=65.0.0
langchain>=0.1.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
langchain-openai>=0.1.0
tiktoken>=0.5.0
